Fixup existing effect handler
